requirements.txt
scikit-image==0.17.2
scipy==1.4.1
tensorflow==1.14.0
matplotlib~=3.2.1
numpy~=1.18.4
nibabel~=3.1.0
opencv-python~=4.2.0.34
Pillow~=7.1.2
imageio~=2.8.0
six~=1.14.0
bunch~=1.0.1
SimpleITK~=1.2.4
scikit-learn~=0.23.0
pynrrd~=0.4.2
